# Ignore trailing spaces in CEF messages