--- a/x-pack/filebeat/processors/decode_cef/cef/parser.go
+++ b/x-pack/filebeat/processors/decode_cef/cef/parser.go
@@ -329,145 +329,177 @@ func (e *Event) unpack(data string) error {
 		case 32:
 			switch data[(p)] {
 			case 32:
-				goto tr54
+				goto tr55
 			case 61:
 				goto tr46
 			case 92:
-				goto tr55
+				goto tr56
+			}
+			if 9 <= data[(p)] && data[(p)] <= 13 {
+				goto tr54
 			}
 			goto tr53
 		case 33:
 			switch data[(p)] {
 			case 32:
-				goto tr56
+				goto tr58
 			case 61:
 				goto tr46
 			case 92:
+				goto tr59
+			}
+			if 9 <= data[(p)] && data[(p)] <= 13 {
 				goto tr57
 			}
 			goto tr48
 		case 34:
 			switch data[(p)] {
 			case 32:
-				goto tr56
+				goto tr58
 			case 61:
 				goto tr46
 			case 92:
-				goto tr57
+				goto tr59
 			case 95:
-				goto tr58
+				goto tr60
 			}
 			switch {
-			case data[(p)] < 65:
-				if 48 <= data[(p)] && data[(p)] <= 57 {
-					goto tr58
+			case data[(p)] < 48:
+				if 9 <= data[(p)] && data[(p)] <= 13 {
+					goto tr57
 				}
-			case data[(p)] > 90:
-				if 97 <= data[(p)] && data[(p)] <= 122 {
-					goto tr58
+			case data[(p)] > 57:
+				switch {
+				case data[(p)] > 90:
+					if 97 <= data[(p)] && data[(p)] <= 122 {
+						goto tr60
+					}
+				case data[(p)] >= 65:
+					goto tr60
 				}
 			default:
-				goto tr58
+				goto tr60
 			}
 			goto tr48
 		case 35:
 			switch data[(p)] {
 			case 32:
-				goto tr56
+				goto tr58
 			case 44:
-				goto tr59
+				goto tr61
 			case 46:
-				goto tr59
+				goto tr61
 			case 61:
-				goto tr60
+				goto tr62
 			case 92:
-				goto tr57
-			case 95:
 				goto tr59
+			case 95:
+				goto tr61
 			}
 			switch {
-			case data[(p)] < 65:
-				if 48 <= data[(p)] && data[(p)] <= 57 {
-					goto tr59
+			case data[(p)] < 48:
+				if 9 <= data[(p)] && data[(p)] <= 13 {
+					goto tr57
 				}
-			case data[(p)] > 93:
-				if 97 <= data[(p)] && data[(p)] <= 122 {
-					goto tr59
+			case data[(p)] > 57:
+				switch {
+				case data[(p)] > 93:
+					if 97 <= data[(p)] && data[(p)] <= 122 {
+						goto tr61
+					}
+				case data[(p)] >= 65:
+					goto tr61
 				}
 			default:
-				goto tr59
+				goto tr61
 			}
 			goto tr48
 		case 36:
 			switch data[(p)] {
 			case 32:
-				goto tr62
+				goto tr65
 			case 61:
 				goto tr46
 			case 92:
-				goto tr63
+				goto tr66
 			}
-			goto tr61
+			if 9 <= data[(p)] && data[(p)] <= 13 {
+				goto tr64
+			}
+			goto tr63
 		case 37:
 			switch data[(p)] {
 			case 32:
-				goto tr64
+				goto tr68
 			case 61:
 				goto tr46
 			case 92:
-				goto tr65
+				goto tr69
+			}
+			if 9 <= data[(p)] && data[(p)] <= 13 {
+				goto tr67
 			}
 			goto tr47
 		case 38:
 			switch data[(p)] {
 			case 32:
-				goto tr64
+				goto tr68
 			case 61:
 				goto tr46
 			case 92:
-				goto tr65
+				goto tr69
 			case 95:
-				goto tr66
+				goto tr70
 			}
 			switch {
-			case data[(p)] < 65:
-				if 48 <= data[(p)] && data[(p)] <= 57 {
-					goto tr66
+			case data[(p)] < 48:
+				if 9 <= data[(p)] && data[(p)] <= 13 {
+					goto tr67
 				}
-			case data[(p)] > 90:
-				if 97 <= data[(p)] && data[(p)] <= 122 {
-					goto tr66
+			case data[(p)] > 57:
+				switch {
+				case data[(p)] > 90:
+					if 97 <= data[(p)] && data[(p)] <= 122 {
+						goto tr70
+					}
+				case data[(p)] >= 65:
+					goto tr70
 				}
 			default:
-				goto tr66
+				goto tr70
 			}
 			goto tr47
 		case 39:
 			switch data[(p)] {
 			case 32:
-				goto tr64
+				goto tr68
 			case 44:
-				goto tr67
+				goto tr71
 			case 46:
-				goto tr67
+				goto tr71
 			case 61:
-				goto tr60
+				goto tr62
 			case 92:
-				goto tr65
+				goto tr69
 			case 95:
-				goto tr67
+				goto tr71
 			}
 			switch {
-			case data[(p)] < 65:
-				if 48 <= data[(p)] && data[(p)] <= 57 {
+			case data[(p)] < 48:
+				if 9 <= data[(p)] && data[(p)] <= 13 {
 					goto tr67
 				}
-			case data[(p)] > 93:
-				if 97 <= data[(p)] && data[(p)] <= 122 {
-					goto tr67
+			case data[(p)] > 57:
+				switch {
+				case data[(p)] > 93:
+					if 97 <= data[(p)] && data[(p)] <= 122 {
+						goto tr71
+					}
+				case data[(p)] >= 65:
+					goto tr71
 				}
 			default:
-				goto tr67
+				goto tr71
 			}
 			goto tr47
 		case 26:
@@ -678,16 +710,16 @@ func (e *Event) unpack(data string) error {
 	tr43:
 		cs = 25
 		goto f0
-	tr65:
+	tr69:
 		cs = 26
 		goto _again
-	tr63:
+	tr66:
 		cs = 26
 		goto f20
-	tr57:
+	tr59:
 		cs = 27
 		goto _again
-	tr55:
+	tr56:
 		cs = 27
 		goto f20
 	tr49:
@@ -708,43 +740,55 @@ func (e *Event) unpack(data string) error {
 	tr45:
 		cs = 32
 		goto f14
+	tr57:
+		cs = 33
+		goto _again
 	tr48:
 		cs = 33
 		goto f16
 	tr53:
 		cs = 33
 		goto f19
-	tr56:
-		cs = 34
-		goto f16
 	tr54:
+		cs = 33
+		goto f20
+	tr58:
 		cs = 34
-		goto f19
-	tr59:
+		goto _again
+	tr55:
+		cs = 34
+		goto f20
+	tr61:
 		cs = 35
 		goto f16
-	tr58:
+	tr60:
 		cs = 35
 		goto f22
-	tr60:
+	tr62:
 		cs = 36
 		goto f14
+	tr67:
+		cs = 37
+		goto _again
 	tr47:
 		cs = 37
 		goto f16
-	tr61:
+	tr63:
 		cs = 37
 		goto f19
 	tr64:
+		cs = 37
+		goto f20
+	tr68:
 		cs = 38
-		goto f16
-	tr62:
+		goto _again
+	tr65:
 		cs = 38
-		goto f19
-	tr67:
+		goto f20
+	tr71:
 		cs = 39
 		goto f16
-	tr66:
+	tr70:
 		cs = 39
 		goto f23
 	tr52:
@@ -830,6 +874,8 @@ func (e *Event) unpack(data string) error {
 		(p)--
 		cs = 28
 		goto _again
+
+		goto _again
 	f17:
 //line cef.rl:87
 
@@ -837,6 +883,8 @@ func (e *Event) unpack(data string) error {
 		// Resume processing at p, the start of the next extension key.
 		p = mark
 		cs = 24
+		goto _again
+
 		goto _again
 	f2:
 //line cef.rl:37
@@ -973,7 +1021,7 @@ func (e *Event) unpack(data string) error {
 					extKey, extValueStart, extValueEnd = "", 0, 0
 				}
 
-//line parser.go:847
+//line parser.go:883
 			}
 		}