--- a/x-pack/filebeat/processors/decode_cef/cef/parser.go
+++ b/x-pack/filebeat/processors/decode_cef/cef/parser.go
@@ -874,8 +874,6 @@ func (e *Event) unpack(data string) error {
 		(p)--
 		cs = 28
 		goto _again
-
-		goto _again
 	f17:
 //line cef.rl:87
 
@@ -883,8 +881,6 @@ func (e *Event) unpack(data string) error {
 		// Resume processing at p, the start of the next extension key.
 		p = mark
 		cs = 24
-		goto _again
-
 		goto _again
 	f2:
 //line cef.rl:37