@adriansr 2020-03-26
@andrewkroh I ran some benchmarks with count=10 and comparing with [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat?tab=doc).

All implementations are compared against the original state machine that doesn't trim.

## Original vs modified SM (this PR)

```
name                      old time/op  new time/op  delta
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef goos:darwin goarch:amd64
ProcessorRun/short_msg-8  5.29µs ± 0%  5.34µs ± 1%  +1.02%  (p=0.000 n=9+9)
ProcessorRun/long_msg-8   65.5µs ± 1%  65.7µs ± 0%    ~     (p=0.165 n=10+10)
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef/cef goos:darwin goarch:amd64
EventUnpack-8             2.17µs ± 1%  2.14µs ± 1%  -1.25%  (p=0.000 n=10+9)
```

##  Original vs strings.TrimRight [[changes]](https://github.com/adriansr/beats/commit/d9c9fc00415540a7458599dc1b5be6739823c17e)

```
name                      old time/op  new time/op  delta
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef goos:darwin goarch:amd64
ProcessorRun/short_msg-8  5.29µs ± 0%  5.36µs ± 1%  +1.29%  (p=0.000 n=9+10)
ProcessorRun/long_msg-8   65.5µs ± 1%  64.3µs ± 1%  -1.94%  (p=0.000 n=10+10)
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef/cef goos:darwin goarch:amd64
EventUnpack-8             2.17µs ± 1%  2.17µs ± 1%    ~     (p=0.388 n=10+9)
```

## Original vs Custom Loop [[changes]](https://github.com/adriansr/beats/commit/9e23121a16be39c47a001a05ab803f941817f812)

```
name                      old time/op  new time/op  delta
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef goos:darwin goarch:amd64
ProcessorRun/short_msg-8  5.29µs ± 0%  5.30µs ± 1%    ~     (p=0.660 n=9+10)
ProcessorRun/long_msg-8   65.5µs ± 1%  64.4µs ± 1%  -1.72%  (p=0.000 n=10+9)
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef/cef goos:darwin goarch:amd64
EventUnpack-8             2.17µs ± 1%  2.08µs ± 1%  -4.36%  (p=0.000 n=10+10)
```

## Original vs Dedicated state-machine [[changes]](https://github.com/adriansr/beats/commit/df38644cabbc047dbb3a9e587e4e122b8a49a38c)
Did this one just for the sake of it.

```
name                      old time/op  new time/op  delta
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef goos:darwin goarch:amd64
ProcessorRun/short_msg-8  5.29µs ± 0%  5.40µs ± 1%  +2.05%  (p=0.000 n=9+10)
ProcessorRun/long_msg-8   65.5µs ± 1%  66.1µs ± 1%  +0.91%  (p=0.001 n=10+10)
pkg:github.com/elastic/beats/v7/x-pack/filebeat/processors/decode_cef/cef goos:darwin goarch:amd64
EventUnpack-8             2.17µs ± 1%  2.16µs ± 1%  -0.61%  (p=0.003 n=10+10)
```